Unet class Model
In unet.py you can find the unet class model in pytorch. It is based on the architecture of the network in the paper " Satellite Image Segmentation for Building Detection using U-Net", given by the prof. It takes in input a RGB image, and returns a binary image of the same dimensions of the input.

In the notebook you may try to play with an instance of the model, and observe its output with no training.